Refactor: Clothes Crawling 및 Controller 구현 #35
1. TopResponse, BottomResponse id 추가
2. Crawling, api 요청 Service 분리
3. findbyClothesGroup 예외처리
