+++ src/main/java/com/seungah/todayclothes/domain/clothes/dto/response/TopResponse.java
@@ -11,11 +11,13 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class TopResponse {
+    private Long id;
     private String itemUrl;
     private String imgUrl;
 
     public static TopResponse of(Top top) {
         return TopResponse.builder()
+                .id(top.getId())
                 .itemUrl(top.getItemUrl())
                 .imgUrl(top.getImgUrl())
                 .build();
