+++ src/main/java/com/seungah/todayclothes/domain/clothes/repository/ClothesGroupRepository.java
@@ -4,7 +4,9 @@ import com.seungah.todayclothes.domain.clothes.entity.ClothesGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
 public interface ClothesGroupRepository extends JpaRepository<ClothesGroup, Long> {
-    ClothesGroup findByGroupNumber(Integer groupNumber);
+    Optional<ClothesGroup> findByGroupNumber(Integer groupNumber);
 }
